fix register
