+++ lib/layers/presentation/pages/register_screen.dart
@@ -10,6 +10,7 @@ class RegisterScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    PageController pageController = PageController();
     return Scaffold(
         backgroundColor: primary,
         body: SafeArea(
@@ -57,66 +58,12 @@ class RegisterScreen extends StatelessWidget {
                   ),
                 ),
                 Container(
-                  padding: EdgeInsets.only(top: 15, left: 34),
-                  child: Text(
-                    'Nama',
-                    style: GoogleFonts.comfortaa(
-                      textStyle: Typo.title.copyWith(
-                        color: secondary,
-                        fontSize: 15,
-                        fontWeight: FontWeight.w700,
-                      ),
-                    ),
+                  height: 320,
+                  child: PageView(
+                    physics: NeverScrollableScrollPhysics(),
+                    controller: pageController,
+                    children: [register(), verif()],
                   ),
-                ),
-                InputLabel(label: 'Masukkan Nama'),
-                Container(
-                  padding: EdgeInsets.only(top: 15, left: 34),
-                  child: Text(
-                    'Email',
-                    style: GoogleFonts.comfortaa(
-                      textStyle: Typo.title.copyWith(
-                        color: secondary,
-                        fontSize: 15,
-                        fontWeight: FontWeight.w700,
-                      ),
-                    ),
-                  ),
-                ),
-                InputLabel(label: 'Masukkan Email'),
-                Container(
-                  padding: EdgeInsets.only(top: 15, left: 34),
-                  child: Text(
-                    'Password',
-                    style: GoogleFonts.comfortaa(
-                      textStyle: Typo.title.copyWith(
-                        color: secondary,
-                        fontSize: 15,
-                        fontWeight: FontWeight.w700,
-                      ),
-                    ),
-                  ),
-                ),
-                InputLabel(
-                  label: 'Masukkan Password',
-                  isOsecure: true,
-                ),
-                Container(
-                  padding: EdgeInsets.only(top: 15, left: 34),
-                  child: Text(
-                    'Konfirmasi Password',
-                    style: GoogleFonts.comfortaa(
-                      textStyle: Typo.title.copyWith(
-                        color: secondary,
-                        fontWeight: FontWeight.w700,
-                        fontSize: 15,
-                      ),
-                    ),
-                  ),
-                ),
-                InputLabel(
-                  label: 'Masukkan Password',
-                  isOsecure: true,
                 ),
                 SizedBox(
                   height: 111,
@@ -137,7 +84,7 @@ class RegisterScreen extends StatelessWidget {
                       child: Container(
                         padding: EdgeInsets.only(top: 12),
                         child: Text(
-                          'Selanjutnya',
+                          'Daftar',
                           textAlign: TextAlign.center,
                           style: GoogleFonts.comfortaa(
                             textStyle: Typo.title.copyWith(
@@ -150,7 +97,13 @@ class RegisterScreen extends StatelessWidget {
                       ),
                     ),
                     onTap: () {
-                      router.pushNamed('verif');
+                      if (pageController.page == 0) {
+                        pageController.animateToPage(1,
+                            duration: const Duration(milliseconds: 500),
+                            curve: Curves.easeInOut);
+                      } else if (pageController.page == 1) {
+                        router.pop();
+                      }
                     },
                   )
                 ])),
@@ -189,7 +142,13 @@ class RegisterScreen extends StatelessWidget {
                         ),
                       ),
                       onTap: () {
-                        router.pushNamed('login');
+                        if (pageController.page == 1) {
+                          pageController.animateToPage(0,
+                              duration: const Duration(milliseconds: 500),
+                              curve: Curves.easeInOut);
+                        } else if (pageController.page == 0) {
+                          router.pop();
+                        }
                       },
                     ),
                   ],
@@ -198,6 +157,146 @@ class RegisterScreen extends StatelessWidget {
             ),
           ),
         ));
+  }
+
+  Widget register() {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Container(
+          padding: EdgeInsets.only(top: 15, left: 34),
+          child: Text(
+            'Nama',
+            style: GoogleFonts.comfortaa(
+              textStyle: Typo.title.copyWith(
+                color: secondary,
+                fontSize: 15,
+                fontWeight: FontWeight.w700,
+              ),
+            ),
+          ),
+        ),
+        InputLabel(label: 'Masukkan Nama'),
+        Container(
+          padding: EdgeInsets.only(top: 15, left: 34),
+          child: Text(
+            'Email',
+            style: GoogleFonts.comfortaa(
+              textStyle: Typo.title.copyWith(
+                color: secondary,
+                fontSize: 15,
+                fontWeight: FontWeight.w700,
+              ),
+            ),
+          ),
+        ),
+        InputLabel(label: 'Masukkan Email'),
+        Container(
+          padding: EdgeInsets.only(top: 15, left: 34),
+          child: Text(
+            'Password',
+            style: GoogleFonts.comfortaa(
+              textStyle: Typo.title.copyWith(
+                color: secondary,
+                fontSize: 15,
+                fontWeight: FontWeight.w700,
+              ),
+            ),
+          ),
+        ),
+        InputLabel(
+          label: 'Masukkan Password',
+          isOsecure: true,
+        ),
+        Container(
+          padding: EdgeInsets.only(top: 15, left: 34),
+          child: Text(
+            'Konfirmasi Password',
+            style: GoogleFonts.comfortaa(
+              textStyle: Typo.title.copyWith(
+                color: secondary,
+                fontWeight: FontWeight.w700,
+                fontSize: 15,
+              ),
+            ),
+          ),
+        ),
+        InputLabel(
+          label: 'Masukkan Password',
+          isOsecure: true,
+        ),
+      ],
+    );
+  }
+
+  Widget verif() {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Container(
+          padding: EdgeInsets.only(top: 15, left: 34),
+          child: Text(
+            'Alamat',
+            style: GoogleFonts.comfortaa(
+              textStyle: Typo.title.copyWith(
+                color: secondary,
+                fontSize: 15,
+                fontWeight: FontWeight.w700,
+              ),
+            ),
+          ),
+        ),
+        InputLabel(label: 'Masukkan Alamat'),
+        Container(
+          padding: EdgeInsets.only(top: 15, left: 34),
+          child: Text(
+            'Nomor Telepon',
+            style: GoogleFonts.comfortaa(
+              textStyle: Typo.title.copyWith(
+                color: secondary,
+                fontSize: 15,
+                fontWeight: FontWeight.w700,
+              ),
+            ),
+          ),
+        ),
+        InputLabel(label: 'Masukkan Nomor Telepon'),
+        Container(
+          padding: EdgeInsets.only(top: 15, left: 34),
+          child: Text(
+            'Gambar Pendukung',
+            style: GoogleFonts.comfortaa(
+              textStyle: Typo.title.copyWith(
+                color: secondary,
+                fontSize: 15,
+                fontWeight: FontWeight.w700,
+              ),
+            ),
+          ),
+        ),
+        InputLabel2(
+          label: 'Pilih Gambar Pendukung',
+          isOsecure: true,
+        ),
+        Container(
+          padding: EdgeInsets.only(top: 15, left: 34),
+          child: Text(
+            'Sertifikat',
+            style: GoogleFonts.comfortaa(
+              textStyle: Typo.title.copyWith(
+                color: secondary,
+                fontWeight: FontWeight.w700,
+                fontSize: 15,
+              ),
+            ),
+          ),
+        ),
+        InputLabel2(
+          label: 'Pilih Sertifikat',
+          isOsecure: true,
+        ),
+      ],
+    );
   }
 }
 
@@ -225,6 +324,55 @@ class InputLabel extends StatelessWidget {
         obscureText: isOsecure ?? false,
         keyboardType: TextInputType.emailAddress,
         decoration: InputDecoration(
+          border: UnderlineInputBorder(),
+          enabledBorder: UnderlineInputBorder(
+            borderSide: BorderSide(color: white!),
+          ),
+          focusedBorder: UnderlineInputBorder(
+            borderSide: BorderSide(color: white!),
+          ),
+          hintText: label,
+          hintStyle: GoogleFonts.comfortaa(
+            textStyle: Typo.title.copyWith(
+              color: white,
+              fontSize: 15,
+              fontWeight: FontWeight.w700,
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class InputLabel2 extends StatelessWidget {
+  InputLabel2({
+    super.key,
+    required this.label,
+    this.isOsecure,
+  });
+  final String? label;
+  bool? isOsecure;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: EdgeInsets.only(left: 34, right: 34),
+      child: TextField(
+        style: GoogleFonts.comfortaa(
+          textStyle: Typo.title.copyWith(
+            color: white,
+            fontSize: 15,
+            fontWeight: FontWeight.w700,
+          ),
+        ),
+        obscureText: isOsecure ?? false,
+        keyboardType: TextInputType.emailAddress,
+        decoration: InputDecoration(
+          icon: Icon(
+            Icons.camera_alt_rounded,
+            color: white,
+          ),
           border: UnderlineInputBorder(),
           enabledBorder: UnderlineInputBorder(
             borderSide: BorderSide(color: white!),
